+++ admin-ui/src/component/my-units/MyUnitRecurringReservation/RecurringReservationDone.tsx
@@ -7,7 +7,7 @@ import styled from "styled-components";
 import { Container } from "hds-react";
 import { H1, H6 } from "common/src/common/typography";
 import { ActionsWrapper } from "./commonStyling";
-import { ReservationList } from "./ReservationsList";
+import ReservationList from "./ReservationsList";
 import withMainMenu from "../../withMainMenu";
 
 const InfoSection = styled.p`
