+++ admin-ui/src/component/my-units/MyUnitRecurringReservation/ReservationsList.test.tsx
@@ -6,7 +6,8 @@ import { render } from "@testing-library/react";
 import { addDays, nextMonday } from "date-fns";
 import { ReservationUnitsReservationUnitReservationStartIntervalChoices } from "common/types/gql-types";
 import { toUIDate } from "common/src/common/util";
-import { generateReservations, ReservationList } from "./ReservationsList";
+import { generateReservations } from "./generateReservations";
+import ReservationList from "./ReservationsList";
 
 const today = new Date();
 const dtoday = Date.UTC(today.getFullYear(), today.getMonth(), today.getDate());
